s3: add provider quirk --s3-might-gzip to fix Cloudflare corrupted on transfer
Before this change, some files were giving this error when downloaded
from Cloudflare

    ERROR corrupted on transfer: sizes differ NNN vs MMM

This is because Cloudflare auto gzips the object when rclone wasn't
expecting it to. (None of the other providers gzip objects without
their being uploaded gzipped).

This patch adds a quirk for Cloudflare to fix the problem and a flag
to control it.

See: https://forum.rclone.org/t/s3-error-corrupted-on-transfer-sizes-differ-nnn-vs-mmm/33694/